requirements.txt
flask
flask-cors
python-dotenv
openai==1.38.0